refactor(components): move styled components
