+++ src/components/RangeSlider/index.tsx
@@ -7,8 +7,7 @@ import {
   SliderThumb,
   SliderMarker,
   SliderMarkerLabel,
-} from './styled'
-
+} from 'components/styled'
 import { useRangeSlider } from 'hooks/useRangeSlider'
 
 import { IRangeMarker, TRangeTuple } from 'types'
@@ -48,7 +47,7 @@ const RangeSlider: React.FC<IRangeSlider> = ({
   return (
     <SliderContainer>
       <SliderRail {...getRailProps()} />
-      <SliderTrack data-testid="range-track" {...getTrackProps()} />
+      <SliderTrack data-testid="range-slider-track" {...getTrackProps()} />
 
       {markers?.map((marker) => {
         const { style, isInRange } = getMarkerProps(marker)
@@ -63,8 +62,8 @@ const RangeSlider: React.FC<IRangeSlider> = ({
         )
       })}
 
-      <SliderThumb data-testid="range-min-thumb" {...getMinHandleProps()} />
-      <SliderThumb data-testid="range-max-thumb" {...getMaxHandleProps()} />
+      <SliderThumb data-testid="range-slider-min-thumb" {...getMinHandleProps()} />
+      <SliderThumb data-testid="range-slider-max-thumb" {...getMaxHandleProps()} />
     </SliderContainer>
   )
 }
